tao-api/requirements.txt
asyncmy
sqlalchemy[asyncio]
fastapi
bittensor
uvicorn
redis[hiredis]
python-decouple
python-multipart
celery[redis]